Started working on polygon layer
In this layer, the capturing area is selected using a polygon.

For finding optimal paths we need to use heuristics. The idea behind heuristics is an assumption that the shortest path will always be parallel to the edge of the polygon.

The paths themselves are built using gnomonic projection which allows us to work with geodesics as with vectors.

The current implementation fails in some cases (I'm not sure which ones, it seem to fail at random), and it's incredibly slow.

I'm about to try to do all calculations in gnomonic projection but not sure if it'll work.

Anyway, this commit also introduces massive refactoring for the new layer to work.

Some misc changes:

- Added polygons support to the DEM parsers.
- GeoTIFFParser now uses point-in-polygon instead of point-in-rectangle in projected coordinates.
- Localized a notification when exporting a layer with no paths. The text now also fits all polygon-based layers.